+++ WorldClass.h
@@ -16,8 +16,8 @@ public:
     ~world();
     int getWeightedTile(const float weights[tileNum]);
     int pickTileFromWeight(int indexi, int indexj);
-    void draw(GamesEngineeringBase::Window& canvas, int _x, int _y);
-    void update(int& mapx, int& mapy);
+    void draw(GamesEngineeringBase::Window& canvas, float _x, float _y);
+    void update(float& mapx, float& mapy);
     void print();
 };
